refactor: 将编辑前须知剥离
